Move package name from com.vaadin to com.vaadin.icons .
